types correction - long to int

+++ IscDbc/Blob.h
@@ -17,10 +17,10 @@ public:
 	virtual void	addRef() = 0;
 	virtual int		release() = 0;
 	virtual void	clear() = 0;
-	virtual void	getBytes (long pos, long length, void *buffer) = 0;
-	virtual void	getBytesW (long pos, long length, void *buffer) = 0;
-	virtual void	getBinary (long pos, long length, void *buffer) = 0;
-	virtual void	getHexString (long pos, long length, void *buffer) = 0;
+	virtual void	getBytes (int pos, int length, void *buffer) = 0;
+	virtual void	getBytesW (int pos, int length, void *buffer) = 0;
+	virtual void	getBinary (int pos, int length, void *buffer) = 0;
+	virtual void	getHexString (int pos, int length, void *buffer) = 0;
 	virtual int		length() = 0;
 	virtual int		getSegmentLength (int pos) = 0;
 	virtual void	putSegment (int length, const char *data, bool copyFlag) = 0;
@@ -28,13 +28,13 @@ public:
 	virtual void	*getSegment (int pos) = 0;
 	virtual void	writeBlob(char * sqldata) = 0;
 	virtual void	writeStreamHexToBlob(char * sqldata) = 0;
-	virtual void	writeBlob(char * sqldata, char *data, long length) = 0;
-	virtual void	writeStringHexToBlob(char * sqldata, char *data, long length) = 0;
+	virtual void	writeBlob(char * sqldata, char *data, int length) = 0;
+	virtual void	writeStringHexToBlob(char * sqldata, char *data, int length) = 0;
 	virtual void	directCreateBlob( char * sqldata ) = 0;
 	virtual void	directOpenBlob(char * sqldata) = 0;
 	virtual bool	directFetchBlob(char *data, int length, int &lengthRead) = 0;
 	virtual bool	directGetSegmentToHexStr( char * bufData, int lenData, int &lenRead ) = 0;
-	virtual void	directWriteBlob( char *data, long length ) = 0;
+	virtual void	directWriteBlob( char *data, int length ) = 0;
 	virtual void	directCloseBlob() = 0;
 	virtual int		getOffset() = 0;
